git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/svn_mutex@1182053 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_repos/replay.c
@@ -136,7 +136,7 @@ struct path_driver_cb_baton
   void *authz_read_baton;
 
   const char *base_path; /* relpath */
-  int base_path_len;
+  size_t base_path_len;
 
   svn_revnum_t low_water_mark;
   /* Stack of active copy operations. */
@@ -372,7 +372,7 @@ path_driver_cb_func(void **dir_baton,
   svn_fs_root_t *source_root = cb->compare_root;
   const char *source_fspath = NULL;
   const char *base_path = cb->base_path;
-  int base_path_len = cb->base_path_len;
+  size_t base_path_len = cb->base_path_len;
 
   *dir_baton = NULL;
